refactor: rename some internal functions, and rename `module.py` to `modules.py`;
