+++ docs/pypots.cli.rst
@@ -1,0 +1,56 @@
+pypots.cli package
+==================
+
+pypots.cli.base module
+----------------------
+
+.. automodule:: pypots.cli.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
+
+pypots.cli.dev module
+---------------------
+
+.. automodule:: pypots.cli.dev
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
+
+pypots.cli.doc module
+---------------------
+
+.. automodule:: pypots.cli.doc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
+
+pypots.cli.env module
+---------------------
+
+.. automodule:: pypots.cli.env
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
+
+pypots.cli.pypots\_cli module
+-----------------------------
+
+.. automodule:: pypots.cli.pypots_cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
+
+Module contents
+---------------
+
+.. automodule:: pypots.cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
