+++ pypots/classification/raindrop/model.py
@@ -25,7 +25,7 @@ from torch.utils.data import DataLoader
 
 from pypots.classification.base import BaseNNClassifier
 from pypots.classification.grud.dataset import DatasetForGRUD
-from pypots.classification.raindrop.module import (
+from pypots.classification.raindrop.modules import (
     PositionalEncoding,
     ObservationPropagation,
 )
@@ -517,7 +517,7 @@ class Raindrop(BaseNNClassifier):
         self.model.eval()  # set the model as eval status to freeze it.
 
         # Step 3: save the model if necessary
-        self.auto_save_model_if_necessary(training_finished=True)
+        self._auto_save_model_if_necessary(training_finished=True)
 
     def classify(self, X: Union[dict, str], file_type: str = "h5py") -> np.ndarray:
         """Classify the input data with the trained model.
