+++ pypots/forecasting/base.py
@@ -198,7 +198,7 @@ class BaseNNForecaster(BaseNNModel, BaseForecaster):
 
                     # save training loss logs into the tensorboard file for every step if in need
                     if self.summary_writer is not None:
-                        self.save_log_into_tb_file(training_step, "training", results)
+                        self._save_log_into_tb_file(training_step, "training", results)
 
                 # mean training loss of the current epoch
                 mean_train_loss = np.mean(epoch_train_loss_collector)
@@ -219,7 +219,7 @@ class BaseNNForecaster(BaseNNModel, BaseForecaster):
                         val_loss_dict = {
                             "imputation_loss": mean_val_loss,
                         }
-                        self.save_log_into_tb_file(epoch, "validating", val_loss_dict)
+                        self._save_log_into_tb_file(epoch, "validating", val_loss_dict)
 
                     logger.info(
                         f"epoch {epoch}: "
