+++ pypots/imputation/base.py
@@ -250,7 +250,7 @@ class BaseNNImputer(BaseNNModel, BaseImputer):
 
                     # save training loss logs into the tensorboard file for every step if in need
                     if self.summary_writer is not None:
-                        self.save_log_into_tb_file(training_step, "training", results)
+                        self._save_log_into_tb_file(training_step, "training", results)
 
                 # mean training loss of the current epoch
                 mean_train_loss = np.mean(epoch_train_loss_collector)
@@ -279,7 +279,7 @@ class BaseNNImputer(BaseNNModel, BaseImputer):
                         val_loss_dict = {
                             "imputation_loss": mean_val_loss,
                         }
-                        self.save_log_into_tb_file(epoch, "validating", val_loss_dict)
+                        self._save_log_into_tb_file(epoch, "validating", val_loss_dict)
 
                     logger.info(
                         f"epoch {epoch}: "
@@ -296,7 +296,7 @@ class BaseNNImputer(BaseNNModel, BaseImputer):
                     self.best_model_dict = self.model.state_dict()
                     self.patience = self.original_patience
                     # save the model if necessary
-                    self.auto_save_model_if_necessary(
+                    self._auto_save_model_if_necessary(
                         training_finished=False,
                         saving_name=f"{self.__class__.__name__}_epoch{epoch}_loss{mean_loss}",
                     )
